requirements.txt
-e git://github.com/gabrielfalcao/lettuce.git@6e7a5566f13f01192d65b1fc010659d88a27a546#egg=lettuce
-e git://github.com/kmmbvnr/django-jenkins.git#egg=django-jenkins
-e git://github.com/jonmer85/django-color-utils.git#egg=color_utils
-e git://github.com/jonmer85/django-basic-feedback.git@10ca27c5824886fabe6a9dd74cf52284b3b34674#egg=django_basic_feedback-dev
-e git://github.com/jonmer85/django-oscar.git@dec47b79a933a014cf01c1f49c00b0bf6df084bc#egg=django-oscar
-e git://github.com/jonmer85/django-oscar-stripe.git#egg=oscar_stripe
-e git://github.com/jonmer85/django-image-cropping.git@bf5963cc8423cf244cff17d32b52624913a7539b#egg=image_cropping
-e git://github.com/fabika/django-seo.git@3fa5eb34de0ed06cfacd4fc401bb0f5e4724e370#egg=DjangoSEO
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
Django==1.7.1
South==0.8.4
Unidecode==0.04.14
Unipath==1.0
argparse==1.2.1
astroid==1.0.0
billiard==3.3.0.18
bleach==1.4.1
boto==2.8.0
celery==3.1.16
coverage==3.6
cssselect==0.9.1
dj-database-url==0.2.1
django-appconf==1.0.1
django-autoslug==1.7.1
django-backbone==0.3.1
django-bleach==0.3.0
django-braces==1.2.2
django-celery==3.1.16
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-dotenv==1.3.0
django-endless-pagination==2.0
django-extensions==1.3.2
django-extra-views==0.6.3
django-fixture-media==0.1.4
django-floppyforms==1.3.0
django-getenv==1.3.1
django-haystack==2.1.0
django-nose==1.2
django-parsley==0.5
django-redis-cache==0.13.0
django-robots==1.0
django-selenium==0.9.7
django-smart-load-tag==0.3.2
django-sslify>=0.2.0
django-storages==1.1.6
django-tables2==0.15.0
django-tinymce==1.5.2
django-treebeard==2.0b2
duvet==0.1.2
easypost==2.0.11
easy-thumbnails==2.2
extras==0.0.3
fuzzywuzzy==0.2
gunicorn==0.17.4
html5lib==0.999
kombu==3.0.21
logilab-common==0.60.0
lxml==3.2.4
Minidetector==0.1
nose==1.3.0
Pillow==2.7.0
phonenumbers==5.9.2
psycopg2==2.4.6
purl==1.0.1
pylint==1.0.0
python-mimeparse==0.1.4
python-subunit==0.0.16
pytz==2013b
raven>=3
redis==2.10.3
selenium==2.43
six==1.3.0
sorl-thumbnail==12.2
sure==1.2.3
testtools==0.9.34
wsgiref==0.1.2
simplejson==3.5.0
